requirements.txt
beautifulsoup4~=4.12.2
click~=8.1.7
lxml~=5.1.0
odspy~=0.1
pandas~=2.1.3
pillow~=10.2.0
PyMuPDF~=1.23.6
PyMuPDFb~=1.23.6
shapely~=2.0.3